Small refactoring to simplify structure and remove code duplication
